requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
psycopg2-binary>=2.9.0
requests>=2.31.0
python-dotenv>=1.0.0
numpy>=1.24.0
# Novas dependências para embeddings reais e NLP
sentence-transformers>=2.2.0
torch>=2.0.0
transformers>=4.30.0
openai>=1.3.0
# Para pré-processamento de texto em português
nltk>=3.8.0
unidecode>=1.3.0
# Para melhorias de performance
scikit-learn>=1.3.0
langchain
langchain-openai
langchain-community
langchain-chroma
chromadb
PyPDF2
python-magic
beautifulsoup4
lxml
tiktoken
pydantic 